Update README.md
Incorporated the list as a table in an expandable section and removed anything related to the .csv file that had been linked to show the list. Incorporated the content from the .md file that had been linked to the acronym resources. Edited other content for clarity. Will incorporate the table content later.